wined3d: Initialise outputs in wined3d_init().
Place all outputs in a struct wined3d so that D3D9 and older
versions of D3D can index into outputs as adapters because
the older versions of D3D treat outputs as adapters. And
slice all outputs to the adapters they belong to so that DXGI
can find outputs within adapters as well.

Signed-off-by: Zhiyi Zhang <zzhang@codeweavers.com>