ddraw: Move device window back to primary monitor if full screen mode is chosen.
wined3d_swapchain_create() use device window to determine which output
to use. And DirectDraw only supports full screen on the primary monitor.
So move the device window back to the primary monitor before creating
swapchain. Otherwise, the secondary output may be used instead of the
primary output.

Fix Binding of Isaac crash when the game is moved to the secondary monitor
and switch to full screen.

Signed-off-by: Zhiyi Zhang <zzhang@codeweavers.com>