comdlg32: Use large toolbar bitmaps for open file dialog when DPI is greater than 144.
Partially fix toolbar bitmaps being too small when DPI is high in open file dialogs. Because there
are only two set of toolbar standard bitmaps, when DPI is even higher, for example, with 250%
scaling, even with IDB_VIEW_LARGE_COLOR bitmaps, the toolbar on the open file dialog still appears
small relatively.

Note that this might break bug for bug compatibility on Wine because the open file dialog on XP
always use IDB_VIEW_SMALL_COLOR regardless of DPI settings. On newer versions of Windows, the open
file dialog is a new design so it doesn't have this bug. So eventually, we may have to redesign
Wine's open file dialog as well.

Signed-off-by: Zhiyi Zhang <zzhang@codeweavers.com>